fix: change price and amount fields from Double to BigDecimal
Changed all monetary fields (price, totalAmount, subtotal) from Double to
BigDecimal to match database DECIMAL(12,2) schema and avoid floating-point
precision issues.

Changes:
- Order.totalAmount: Double -> BigDecimal
- OrderItem.price, subtotal: Double -> BigDecimal
- ProductVariant.price: Double -> BigDecimal
- CartItem.getSubtotal(): double -> BigDecimal
- All related DTOs (OrderDtos, CartDtos, VariantCreateRequest, ProductUpdateDtos)
- OrderServiceImpl: Updated price calculations to use BigDecimal arithmetic
- DataSeeder: Updated variant creation to use BigDecimal

This fixes Hibernate schema validation error:
"wrong column type encountered in column [total_amount] in table [order];
found [numeric (Types#NUMERIC)], but expecting [float(53) (Types#FLOAT)]"

+++ src/main/java/kg/bishkek/fucent/fusent/dto/CartDtos.java
@@ -3,6 +3,7 @@ package kg.bishkek.fucent.fusent.dto;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
@@ -26,9 +27,9 @@ public class CartDtos {
         String productImage,
         UUID shopId,
         String shopName,
-        Double price,
+        BigDecimal price,
         Integer qty,
-        Double subtotal,
+        BigDecimal subtotal,
         Integer stockQty,
         Instant addedAt
     ) {}
@@ -38,13 +39,13 @@ public class CartDtos {
         UUID userId,
         List<CartItemResponse> items,
         Integer totalItems,
-        Double totalAmount,
+        BigDecimal totalAmount,
         Instant createdAt,
         Instant updatedAt
     ) {}
 
     public record CartSummary(
         Integer totalItems,
-        Double totalAmount
+        BigDecimal totalAmount
     ) {}
 }
