+++ src/main/java/kg/bishkek/fucent/fusent/dto/VariantCreateRequest.java
@@ -3,6 +3,7 @@ package kg.bishkek.fucent.fusent.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 
@@ -11,6 +12,6 @@ public record VariantCreateRequest(
         @NotBlank String sku,
         String barcode,
         String attributesJson,
-        @NotNull Double price,
+        @NotNull BigDecimal price,
         @NotNull Integer stockQty
 ) {}
